refactor: move code to cli/run
